pyproject.toml
[project]
name = "TravelBotDC"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.4",
    "asyncpg>=0.31.0",
    "discord-py>=2.7.1",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.14",
    "python-dotenv>=1.2.2",
]
